frontend/api/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
drf-nested-routers==0.9.0
django-guardian==1.4.9
django-crispy-forms==1.7.0
django-filter==1.1.0
Markdown==2.6.11
djangorestframework-jwt==1.11.0
psycopg2==2.7.3.2
gunicorn==19.7.1
django-cors-headers==2.1.0
django-rest-auth==0.9.2
django-allauth==0.34.0